Remove unused context manager methods
`__enter__()` and `__exit__()` on `SpanWrapper` aren't needed because
the same methods on the base class `opentracing.Span` do what we need
them to do when `SpanWrapper` is used as a context manager.

`__enter__()` and `__exit__()` on `ScopeWrapper` aren't needed
because we use the `@contextmanager` decorator on
`start_active_span(). When the `with start_active_span()` block ends,
the `with` block in the *implementation* of `start_active_span()`
ends as well, which in turn makes the `with` block in OpenTelemetry's
`start_span()` end, which makes the `finally` block in `use_span()`
run, which calls `end()` on the OpenTelemetry span.