Raise exception on invalid parent type
According to the OpenTracing API, `child_of` can only be of type
`Span` or `SpanContext`. In the shim these are represented as
`SpanWrapper` and `SpanContextWrapper`, respectively. In case we
receive a different type in `child_of`, it's safer to bail out than
to try to set an unexpected object as the span's parent.